Rename : modify blog Template folder structure
